Enable Kokkos OpenACC build
